requirements.txt
crawl4ai
nest_asyncio
requests
supabase
playwright
google-generativeai
duckduckgo-search
python-dotenv
pytest
genai
openai
tiktoken>=0.4.0